+++ src/database/migrations/2006_01_01_000030_projects_create_table_billing.php
@@ -14,36 +14,44 @@ class ProjectsCreateTableBilling extends Migration {
     {
         if(!Schema::hasTable('006_092_billing'))
         {
-            Schema::create('006_092_todo', function (Blueprint $table) {
+            Schema::create('006_092_billing', function (Blueprint $table) {
                 $table->engine = 'InnoDB';
                 $table->increments('id_092')->unsigned();
 
                 // tarea de donde procede este registro
                 $table->integer('todo_id_092')->unsigned();
 
-                // customer
-                $table->integer('customer_id_092')->unsigned();
-                $table->string('customer_name_092');
-
                 // developer / usuario que se la ha asignado la tarea
-                $table->integer('developer_id_092')->unsigned();
-                $table->string('name_092');
+                $table->integer('developer_id_092')->unsigned()->nullable();
+                $table->integer('developer_name_092')->nullable();
+
+                // cliente al que se le realiza la tarea
+                $table->integer('customer_id_092')->unsigned()->nullable();
+                $table->string('customer_name_092')->nullable();
+
+                // descripción de la tarea
+                $table->string('title_092');
                 $table->text('description_092');
+
+                // precio de la tarea, en el caso de haberse definido
+                $table->decimal('price_092', 10, 2)->nullable();
+
+                // fecha de petición de la tarea
+                $table->integer('request_date_092')->unsigned();
+                $table->string('request_date_text_092');
+
+                // fechas de entrega
+                $table->integer('end_date_092')->unsigned()->nullable();
+                $table->string('end_date_text_092')->nullable();
 
                 // número de horas realizadas en esta tarea
                 $table->decimal('hours_092', 8, 2);
 
-                // precio de la tarea, en el cado de haberse definido
-                $table->decimal('price_092', 10, 2)->nullable();
-
-                // fechas de petición de la tarea y fecha de entrega
-                $table->integer('request_date_092')->unsigned();
-                $table->string('request_date_text_092');
-                $table->integer('end_date_092')->unsigned()->nullable();
-                $table->string('end_date_text_092')->nullable();
-
                 // si está facturado o no
                 $table->tinyInteger('invoiced_092')->unsigned();
+
+                $table->foreign('developer_id_092', 'fk01_006_092_billing')->references('id_010')->on('001_010_user')
+                    ->onDelete('set null')->onUpdate('cascade');
             });
         }
     }
@@ -55,6 +63,9 @@ class ProjectsCreateTableBilling extends Migration {
      */
     public function down()
     {
-        Schema::drop('006_092_billing');
+        if (Schema::hasTable('006_092_billing'))
+        {
+            Schema::drop('006_092_billing');
+        }
     }
 }
